Add Death to frame data and WIP death animation
Death animation is a stripped version of PlayerDeadBody (GhostDeadBody).
The player's death state is sent every frame and it is up to the clients
to notice the change and trigger the animation. Due to the player
becoming visible on an earlier frame than position data starts to arrive
again, GhostDeadBody removes the Ghost so that it does not render until
after the respawn.